OpenCV Basics - 21 - Aruco Marker Tracking
In this final episode we finish being able to track Aruco markers and draw their axes to the screen using C++ in Xcode. I show you a camera calibration, loading the matrix and then using it to properly detect the markers and display the axes.
I hope you have enjoyed the series.

- Correções no Save, Load, e no CalibragemiMac2017.txt
- Implementação do metodo de reconhecimento de múltiplos marcadores.